move some misc functions into spherical
