+++ code/src/stf/include/interface/stf/math/spherical.hpp
@@ -40,6 +40,18 @@ namespace stf::math
     }
 
     template<typename T>
+    inline vec2<T> unit_vector(T const theta)
+    {
+        return vec2<T>(std::cos(theta), std::sin(theta));
+    }
+
+    template<typename T>
+    inline vec3<T> unit_vector(T const theta, T const phi)
+    {
+        return vec3<T>(std::cos(theta) * std::sin(phi), std::sin(theta) * std::sin(phi), std::cos(phi));
+    }
+
+    template<typename T>
     inline math::vec3<T> to_euclidean(T const radius, T const theta, T const phi)
     {
         return radius * unit_vector(theta, phi);
