+++ code/tests/stf/cpp/stf/math/spherical_tests.cpp
@@ -56,4 +56,19 @@ namespace stf::math
         for (scaffolding::spherical::closest_equiv_angle<float> const& test : tests) { scaffolding::spherical::verify(test); }
     }
 
+    TEST(spherical, unit_vector)
+    {
+        std::vector<scaffolding::spherical::unit_vec2<float>> tests = 
+        {
+            { 0, stff::vec2(1, 0) },
+            { stff::constants::half_pi, stff::vec2(0, 1) },
+            { stff::constants::pi, stff::vec2(-1, 0) },
+            { 3 * stff::constants::half_pi, stff::vec2(0, -1) },
+            { stff::constants::tau, stff::vec2(1, 0) },
+            { stff::constants::quarter_pi, stff::vec2(stff::constants::sqrt_two_inv, stff::constants::sqrt_two_inv) },
+        };
+
+        for (scaffolding::spherical::unit_vec2<float> const& test : tests) { scaffolding::spherical::verify(test); }
+    }
+
 } // stf::math
