+++ code/tests/stf/include/private/stf/math/scaffolding/spherical.hpp
@@ -50,4 +50,17 @@ namespace stf::math::scaffolding::spherical
         ASSERT_NEAR(test.expected, math::closest_equiv_angle(test.phi, test.theta), stff::constants::tol) << "Failed to compute closest equivalent angle";
     }
 
+    template<typename T>
+    struct unit_vec2
+    {
+        T const theta;
+        math::vec<T, 2> const expected;
+    };
+
+    template<typename T>
+    void verify(unit_vec2<T> const& test)
+    {
+        ASSERT_EQ(test.expected, math::unit_vector(test.theta)) << "Failed to compute unit vector";
+    }
+
 } // stf::math::scaffolding::spherical
